Added SearchBar component
