+++ package.json
@@ -19,8 +19,6 @@
     "cors": "^2.8.5",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
-    "flowbite": "^1.6.3",
-    "flowbite-react": "^0.3.8",
     "nodemon": "^2.0.20",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
